update test files before separating them
